# openapi.yaml
openapi: 3.0.3
info:
  title: Contractual Liquidity API
  version: 0.3.0
  description: Django/DRF API for contractual cash-flow runs. Decimal amounts are
    strings. Runs are private to their submitting user.
paths:
...
  /api/v1/runs/{id}:
    get:
      operationId: runs_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - runs
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
          description: ''
...
  /api/v1/runs/{id}/input:
    get:
      operationId: runs_input_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-fA-F-]{36}$
        required: true
      tags:
      - runs
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunInput'
          description: ''
...
components:
  schemas:
...
    RunDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        progress:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        error:
          type: string
        as_of_date:
          type: string
          format: date
        entity:
          type: string
          readOnly: true
        entity_name:
          type: string
          readOnly: true
        is_mock:
          type: boolean
          readOnly: true
        engine_version:
          type: string
          maxLength: 32
        started:
          type: string
          format: date-time
          nullable: true
        finished:
          type: string
          format: date-time
          nullable: true
        heartbeat:
          type: string
          format: date-time
          nullable: true
        input_hash:
          type: string
          maxLength: 64
        result:
          readOnly: true
      required:
      - as_of_date
      - created
      - engine_version
      - entity
      - entity_name
      - id
      - input_hash
      - is_mock
      - result
    RunInput:
      type: object
      properties:
        entity:
          type: string
          maxLength: 64
          pattern: ^[-a-zA-Z0-9_]+$
        as_of_date:
          type: string
          format: date
        bucket_days:
          type: array
          items:
            type: integer
            maximum: 36500
            minimum: 1
          default:
          - 1
          - 7
          - 14
          - 30
          - 60
          - 90
          - 180
          - 270
          - 365
          - 730
          - 1095
          - 1825
          maxItems: 30
          minItems: 1
        contracts:
          type: array
          items: {}
          description: Contract objects. Individual invalid/unsupported contracts
            appear in run exceptions; see sample endpoint and API guide for fields.
          maxItems: 2000
          minItems: 1
      required:
      - as_of_date
      - contracts
      - entity
...
    StatusEnum:
      enum:
      - queued
      - running
      - completed
      - completed_with_exceptions
      - failed_validation
      - failed
      - interrupted
      type: string
      description: |-
        * `queued` - Queued
        * `running` - Running
        * `completed` - Completed
        * `completed_with_exceptions` - Completed with exceptions
        * `failed_validation` - Failed validation
        * `failed` - Failed
        * `interrupted` - Interrupted
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"